fix:connector filter conf modal when default filter chosen
