+++ frontend/app/(main)/workspace/connectors/utils/sync-filter-save-guards.ts
@@ -6,26 +6,61 @@ import type { FilterSchemaField } from '../types';
 
 type FilterRow = { operator?: string; value?: unknown };
 
-function hasMeaningfulSyncFilterValue(
-  field: { filterType?: string },
-  f: FilterRow | undefined
-): boolean {
-  if (!f?.operator) return false;
-  if (field.filterType === 'boolean') return true;
-  if (Array.isArray(f.value)) return f.value.length > 0;
-  if (
-    field.filterType === 'datetime' &&
-    f.value &&
-    typeof f.value === 'object' &&
-    !Array.isArray(f.value)
-  ) {
-    const d = f.value as { start?: string | number | null; end?: string | number | null };
-    return (
-      (d.start != null && d.start !== '') ||
-      (d.end != null && d.end !== '')
-    );
+function isFilterRow(raw: unknown): raw is FilterRow {
+  return (
+    typeof raw === 'object' &&
+    raw !== null &&
+    'operator' in raw &&
+    typeof (raw as FilterRow).operator === 'string'
+  );
+}
+
+function isListLikeFilterField(field: FilterSchemaField): boolean {
+  if (field.fieldType === 'MULTISELECT' || field.fieldType === 'TAGS') return true;
+  const ft = String(field.filterType ?? '').toLowerCase();
+  return ft === 'list' || ft === 'multiselect';
+}
+
+/**
+ * Whether a filter row counts as configured (UI active state, summary chips, save guards).
+ * Must stay aligned with default seeding in filters-section.tsx.
+ */
+export function isMeaningfulFilterRow(field: FilterSchemaField, raw: unknown): boolean {
+  if (raw === undefined || raw === null) return false;
+  if (!isFilterRow(raw) || !raw.operator?.trim()) return false;
+  const row = raw;
+  const ft = String(field.filterType ?? '').toLowerCase();
+  if (ft === 'list' || ft === 'multiselect' || isListLikeFilterField(field)) {
+    if (Array.isArray(row.value)) {
+      if (row.value.length > 0 || field.defaultOperator) {
+        return true;
+      }
+    }
+    return false;
   }
-  return f.value !== undefined && f.value !== null && f.value !== '';
+  if (ft === 'boolean') {
+    return row.value === true || row.value === false;
+  }
+  if (ft === 'datetime') {
+    const op = row.operator.toLowerCase();
+    if (op.startsWith('last_')) return true;
+    const v = row.value as { start?: unknown; end?: unknown } | null;
+    if (!v || typeof v !== 'object') return false;
+    const startOk =
+      v.start !== undefined &&
+      v.start !== null &&
+      v.start !== '' &&
+      !(typeof v.start === 'number' && (v.start <= 0 || Number.isNaN(v.start)));
+    const endOk =
+      v.end !== undefined &&
+      v.end !== null &&
+      v.end !== '' &&
+      !(typeof v.end === 'number' && (v.end <= 0 || Number.isNaN(v.end)));
+    return startOk || endOk;
+  }
+  if (row.value === '' || row.value === undefined || row.value === null) return false;
+  if (typeof row.value === 'string' && !row.value.trim()) return false;
+  return true;
 }
 
 /**
@@ -36,10 +71,7 @@ export function hasAnySyncFiltersSelected(
   syncFormValues: Record<string, unknown>
 ): boolean {
   if (!syncFields?.length) return false;
-  return syncFields.some((field) => {
-    const row = syncFormValues[field.name] as FilterRow | undefined;
-    return row ? hasMeaningfulSyncFilterValue(field, row) : false;
-  });
+  return syncFields.some((field) => isMeaningfulFilterRow(field, syncFormValues[field.name]));
 }
 
 /** Manual indexing toggle lives under indexing filters as `enable_manual_sync`. */
